pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "infra-monitor"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "psutil>=5.9.8",
    "rich>=13.7.1",
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "prometheus-client>=0.20.0",
]

[project.scripts]
infra-monitor = "src.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
